Move search component from events to shared
